// src/Models/EntityBankSetup/table1.ts
export class Table1 {
    EntityBankSetupId: string;
    EntityId: string;
    BankId: string;
    IsFileFormatXml: boolean;
    IsFileFormatExcel: boolean;
    IsFileFormatCsv: boolean;
    IsFileSendDaily: boolean;
    IsFileSendWeekly: boolean;
    IsFileSendMonthly: boolean;
    IsAdhoc: boolean;
    TimeDuration: string;
}